+++ about.html
@@ -42,13 +42,11 @@
 
     <main>
         <section id="aboutSection" class="about-welcome-section">
-            <div class="stars-container">
                 <img class="stars star-1" src="images/background/star-1.png" alt="">
                 <img class="stars star-2" src="images/icons/icon_messages.png" alt="">
                 <img class="stars star-3" src="images/background/star-3.png" alt="">
                 <img class="stars star-4" src="images/background/star-4.png" alt="">
                 <img class="stars star-5" src="images/background/star-5.png" alt="">
-            </div>
             <article class="about-welcome-container">
                 <h1 class="title">"Hello! Sharing world"</h1>
                 <h2 class="subtitle">Creative Ideas Yaz Fortozo 2022</h2>
@@ -100,11 +98,14 @@
                 <img src="images/partners/google.png" alt="Google is a partner">
             </div>
         </article>
-        <article class="rights-container">
-            <img src="images/logo2.png" alt="Made by company">
-            <div class="info-container">
-                <p class="text-1">2022 Yaz Fortozo. Some Rights Reserved.</p>
-                <p class="text-2">Lorem ipsum dolor sit, amet consectetur adipisicing elit. Eligendi exercitationem aliquid voluptatum delectus temporibus error! Id praesentium recusandae, quibusdam omnis magnam cum eos. Quae sit assumenda accusamus hic itaque dolorem.</p>
+        <article class="rights-container row">
+            <img class="rights-container col-12-base col-12-small col-5-desktop" src="images/logo2.png" alt="Made by company">
+            <div class="info-container col-12-base col-12-small col-5-desktop">
+                <p class="text-1">Yaz Fortozo 2022. Acknowledgments.</p>
+                <p class="text-2">To my mentor <strong><a target="_blank" rel="noopener" href="https://github.com/VitorGuedesMadeira" >Vitor Guades Madeira</a></strong> for share with me those subtle but importat clues and ideas whose energy leads me to surpass my capabilities, the title looks really amazing with that fire effect behind.
+                    <br><br>To my mentor friend's <strong><a target="_blank" rel="noopener" href="https://github.com/karlavdelgadof" >Karla Delgado</a></strong> who even without know me to much share her page and draws me through the hazy moments, and that beautiful effect when the menu appear sliding from right to left made me apply it into this project.
+                    <br><br>To <strong><a target="_blank" rel="noopener" href="https://www.youtube.com/channel/UCcHF0S6tex06QUtGQDW_isQ">Yaz Fortozo Monroy</a></strong> for allow me to use their logo, brand, name, and business idea. For helped me to create that wonderful patterns used in the welcome section and more over for edit and change all the 20 text of the clients with a unic name professions and description.
+                </p>
             </div>
         </article>
