+++ sass/2_base/_config.scss
@@ -13,11 +13,11 @@
 }
 
 html {
-  font-size: 10px;
+  font-size: 12px;
   scroll-behavior: smooth;
 
   @include mixin.base {
-    font-size: 10px;
+    font-size: 12px;
   }
 
   @include mixin.small {
@@ -56,6 +56,7 @@ ul {
 a {
   text-decoration: none;
   font-family: v.$font-lato;
+  word-wrap: break-word;
 }
 
 ::before {
