+++ sass/3_layout/_about-main.scss
@@ -7,21 +7,21 @@ $top-mobil: 70px;
 .about-welcome-section {
   background: url('#{v.$path-images}/background/about.jpg') no-repeat;
   background-size: cover;
-  height: 100vh;
+  min-height: 100vh;
   display: flex;
   flex-direction: column;
   position: relative;
+  overflow-x: hidden;
 
   & .about-welcome-container {
     display: flex;
     flex-direction: column;
-    position: absolute;
     align-items: center;
     align-self: center;
-    top: $top-mobil;
-    max-width: 1000px;
+    margin-top: $top-mobil;
     width: 95%;
     text-align: center;
+    z-index: 1;
 
     & .title {
       font-size: f.get-font-size('small');
@@ -42,12 +42,12 @@ $top-mobil: 70px;
       font-family: v.$font-lato;
       font-size: f.get-font-size('smaller');
       font-weight: f.get-font-weight('bold');
-      line-height: 30px;
+      line-height: 1.5em;
       background-color: transparentize($color: f.get-color('natural'), $amount: 0.1%);
       border: 1px solid f.get-color('secundary');
       box-shadow: 5px 5px 10px 5px f.get-color('secundary'), -1px -1px 5px 2px f.get-color('secundary'),;
       margin: 30px 0;
-      padding: 20px;
+      padding: 5%;
     }
 
     & .text-2 {
@@ -55,7 +55,7 @@ $top-mobil: 70px;
       font-size: f.get-font-size('small');
       font-weight: f.get-font-weight('bold');
       color: black;
-      margin: 70px 10px 20px 10px;
+      margin: 5% 10px 5% 10px;
     }
 
     & a {
@@ -63,6 +63,8 @@ $top-mobil: 70px;
       font-weight: f.get-font-weight('bold');
       text-decoration: underline;
       transition: all 0.3s linear;
+      margin: 5% 5%;
+      width: 100%;
 
       &:hover {
         color: f.get-color('accent');
@@ -77,7 +79,7 @@ $top-mobil: 70px;
     }
 
     & .about-welcome-container {
-      top: $top-mobil + v.$upper-bar-height + v.$low-bar-height;
+      margin-top: $top-mobil + v.$upper-bar-height + v.$low-bar-height;
     }
   }
 }
